Player variable fix and merged Yggdrasil & Utils to Skript
Fixed issues with player variables throwing errors when set.

The Yggdrasil used in the real Skript 2.2 wasn't the same as the one on
Github. I added the internal UUID serializer in the Yggdrasil of Skript
2.2 to the one I got from Github. Which fixed the issue.

Merged Yggdrasil & Utils projects to the Skript project. (I had to,
because of some reason).