Small patch.
Called V9b because I added a new option for disabling object cannot be
saved warning,  "... i.e cannot be saved...".
Like when setting a variable to a inventory it will say that it cannot
be saved. Those messages can be disabled.

Also fixed a small error when Skript checks for new Skript Fixes
versions.