This isn't exactly good news.
I have figured out that I do not have time to finish the new parser.
Thus, it must go, as it is slowing down other development as long it
is not finished.

I have stored current state of things on branch "newparser".
If someone wishes to finish my work, please open issue.