Adds Damage Event call to the Damage Effects
Currently, the e.damage function does not call the Damage Event, allowing users to bypass any cancellations of the events that they thought that they have in place. A simple test for this is as follows;

```
on damage:
    cancel event
 command /hurtme [<text>]:
    trigger:
        damage player by 5
```

From my testing, the player will still be damaged, despite the damage event being cancelled. This change will force Skript to respect the cancellation of the event, prior to the damage being dealt.