Remove legacy projectile code (#3713)
* Remove legacy projectile code

Njol probably added them when the parameter type of setShooter method changed to support both old and new method signatures. However, this not necessary anymore and reflection is slower, so I made them direct method calls. The shooter parameter of the ProjectileUtils#setShooter can also be changed to ProjectileSource from Object, but this is a public method and that changes the method signature, breaking binary compatibility (although I didn't saw any add-ons use it.)

* Remove the class altogether instead

The p != null ? p.getShooter() : null statements are there to preserve the original methods behaviour. Not sure if they are required.

Added a local variable in ExprAttacker to avoid confusing repeated casts. The edbee is short for EntityDamageByEntityEvent.

* Fix build

Also remove one of the null checks since instanceof already checks for null.