Ask for player's groups off the main thread. (#6192)
* Update Vault to 1.7.3

This version should support 1.13 to 1.17+ according to the Spigot page. I'm not sure what this actually adds but it's usually a good idea to keep dependencies updated and all 1.7.X versions of Vault are supposed to be compatible. :)

* Do group request from another thread.

This change is (basically) pointless -- we still block the main thread while we wait since we absolutely need to, but it prevents the errors from overzealous permissions plugins.

* Invert parse tag.

---------

Co-authored-by: sovdee <10354869+sovdeeth@users.noreply.github.com>