1.20.5 Support (#6617)
* Internal support for aliases item components

* Update aliases submodule

It now targets to the 1.20.5 support branch

* Update isRunningMethod check

* Add ability to opt out of deduplication

This is useful for items that have NBT that is not represented by an ItemMeta implementation (e.g. block entity data)

* Update skript-aliases

* Update skript-aliases

* Update skript-aliases

* Update Gradle to 8.7

* Update Paper to 1.20.6

Also updates Java from 17 to 21

* Fix latest env for testing

* Fix AreaEffectCloudApplyEvent event value

* Fix vehicle dismount events

* Fix potions

* Get visual effects building

Removed ones do not currently work on new versions...

* Fix unused AbstractList implementation

* Fix remaining Java 17 references

* Fix 1.17 testing

* Add missing PotionDataUtils check

* Add missing lang entries

* Rework Test Environments

We cannot use EasyMock on Java 21 as we must stay on an older version (see https://github.com/SkriptLang/Skript/pull/6204#discussion_r1405302009)
Java 21 will be supported for regular testing. A 1.20.4 environment remains for Java 17

* Update workflow names

---------

Co-authored-by: Moderocky <admin@moderocky.com>